requirements.txt
pandas>=2.0
openpyxl>=3.1
tabulate>=0.9
dash>=2.16
gspread>=6.0
google-auth>=2.0
